refactor NodeCard and NodeInner to get NodeShell
